+++ dotnet/src/webdriver/DevTools/v117/V117Target.cs
@@ -1,4 +1,4 @@
-// <copyright file="V114Target.cs" company="WebDriver Committers">
+// <copyright file="V117Target.cs" company="WebDriver Committers">
 // Licensed to the Software Freedom Conservancy (SFC) under one
 // or more contributor license agreements. See the NOTICE file
 // distributed with this work for additional information
@@ -21,22 +21,22 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Text;
 using System.Threading.Tasks;
-using OpenQA.Selenium.DevTools.V114.Target;
+using OpenQA.Selenium.DevTools.V117.Target;
 
-namespace OpenQA.Selenium.DevTools.V114
+namespace OpenQA.Selenium.DevTools.V117
 {
     /// <summary>
-    /// Class providing functionality for manipulating targets for version 114 of the DevTools Protocol
+    /// Class providing functionality for manipulating targets for version 117 of the DevTools Protocol
     /// </summary>
-    public class V114Target : DevTools.Target
+    public class V117Target : DevTools.Target
     {
         private TargetAdapter adapter;
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="V114Target"/> class.
+        /// Initializes a new instance of the <see cref="V117Target"/> class.
         /// </summary>
         /// <param name="adapter">The adapter for the Target domain.</param>
-        public V114Target(TargetAdapter adapter)
+        public V117Target(TargetAdapter adapter)
         {
             this.adapter = adapter;
             adapter.DetachedFromTarget += OnDetachedFromTarget;