+++ dotnet/test/common/DevTools/DevToolsSecurityTest.cs
@@ -6,7 +6,7 @@ using OpenQA.Selenium.Environment;
 
 namespace OpenQA.Selenium.DevTools
 {
-    using CurrentCdpVersion = V116;
+    using CurrentCdpVersion = V117;
 
     [TestFixture]
     public class DevToolsSecurityTest : DevToolsTestFixture