+++ py/selenium/webdriver/chromium/service.py
@@ -15,9 +15,7 @@
 # specific language governing permissions and limitations
 # under the License.
 import typing
-import warnings
 
-from selenium.common import InvalidArgumentException
 from selenium.types import SubprocessStdAlias
 from selenium.webdriver.common import service
 
@@ -29,7 +27,7 @@ class ChromiumService(service.Service):
     :param executable_path: install path of the executable.
     :param port: Port for the service to run on, defaults to 0 where the operating system will decide.
     :param service_args: (Optional) List of args to be passed to the subprocess when launching the executable.
-    :param log_path: (Optional) String to be passed to the executable as `--log-path`.
+    :param log_output: (Optional) int representation of STDOUT/DEVNULL, any IO instance or String path to file.
     :param env: (Optional) Mapping of environment variables for the new process, defaults to `os.environ`.
     """
 
@@ -38,24 +36,17 @@ class ChromiumService(service.Service):
         executable_path: str = None,
         port: int = 0,
         service_args: typing.Optional[typing.List[str]] = None,
-        log_path: typing.Optional[str] = None,
         log_output: SubprocessStdAlias = None,
         env: typing.Optional[typing.Mapping[str, str]] = None,
         **kwargs,
     ) -> None:
         self.service_args = service_args or []
-        self.log_output = log_output
-        if log_path is not None:
-            warnings.warn("log_path has been deprecated, please use log_output", DeprecationWarning, stacklevel=2)
-            self.log_output = log_path
 
-        if "--append-log" in self.service_args or "--readable-timestamp" in self.service_args:
-            if isinstance(self.log_output, str):
-                self.service_args.append(f"--log-path={self.log_output}")
-                self.log_output = None
-            else:
-                msg = "Appending logs and readable timestamps require log output to be a string representing file path"
-                raise InvalidArgumentException(msg)
+        if isinstance(log_output, str):
+            self.service_args.append(f"--log-path={log_output}")
+            self.log_output = None
+        else:
+            self.log_output = log_output
 
         super().__init__(
             executable=executable_path,