+++ rb/spec/integration/selenium/webdriver/action_builder_spec.rb
@@ -382,8 +382,10 @@ module Selenium
           iframe = driver.find_element(tag_name: 'iframe')
           driver.switch_to.frame(iframe)
           checkbox = driver.find_element(name: 'scroll_checkbox')
-          sleep 0.5
-          expect(in_viewport?(checkbox)).to be true
+
+          expect {
+            wait.until { in_viewport?(checkbox) }
+          }.not_to raise_error
         end
 
         it 'raises MoveTargetOutOfBoundsError when origin offset is out of viewport',